# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.categoryRequest:
    properties:
      color:
        example: '#f59e0b'
        type: string
      icon:
        example: "\U0001F354"
        type: string
      name:
        example: Alimentación
        type: string
      type:
        example: expense
        type: string
    type: object
  handler.categoryResponse:
    properties:
      color:
        example: '#f59e0b'
        type: string
      created_at:
        type: string
      icon:
        example: "\U0001F354"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Alimentación
        type: string
      type:
        example: expense
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      error:
        example: invalid_credentials
        type: string
    type: object
...
  handler.transactionRequest:
    properties:
      account_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      amount:
        example: 50
        type: number
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      date:
        example: "2025-05-25"
        type: string
      description:
        example: Almuerzo
        type: string
      type:
        example: expense
        type: string
    type: object
  handler.transactionResponse:
    properties:
      account_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      account_name:
        example: Cuenta corriente
        type: string
      amount:
        example: 50
        type: number
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      category_name:
        example: Alimentación
        type: string
      created_at:
        type: string
      date:
        type: string
      description:
        example: Almuerzo
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      type:
        example: expense
        type: string
      updated_at:
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST para gestión de gastos personales
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Devuelve todas las categorías del usuario autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.categoryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Listar categorías
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Crea una nueva categoría de tipo income o expense
      parameters:
      - description: Datos de la categoría
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.categoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.categoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Crear categoría
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Elimina una categoría del usuario autenticado
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar categoría
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Actualiza nombre, ícono y color de una categoría
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.categoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.categoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar categoría
      tags:
      - Categories
...
  /transactions:
    get:
      description: Devuelve transacciones del usuario con filtros opcionales de fecha,
        tipo, cuenta y categoría
      parameters:
      - description: Fecha inicio (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Fecha fin (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: 'Tipo: income o expense'
        in: query
        name: type
        type: string
      - description: ID de cuenta
        in: query
        name: account_id
        type: string
      - description: ID de categoría
        in: query
        name: category_id
        type: string
      - description: 'Orden: asc o desc (por defecto desc)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.transactionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Listar transacciones
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Crea una nueva transacción y ajusta el saldo de la cuenta si se
        especifica
      parameters:
      - description: Datos de la transacción
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.transactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.transactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Crear transacción
      tags:
      - Transactions
  /transactions/{id}:
    delete:
      description: Elimina una transacción y revierte el saldo de la cuenta asociada
      parameters:
      - description: ID de la transacción
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar transacción
      tags:
      - Transactions
    put:
      consumes:
      - application/json
      description: Actualiza una transacción y recalcula el saldo de la cuenta afectada
      parameters:
      - description: ID de la transacción
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.transactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.transactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar transacción
      tags:
      - Transactions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"